Add some styles
