+++ index.html
@@ -188,6 +188,7 @@
                   <option value="frame" data-text="frame">Frame</option>
                   <option value="subtle" data-text="subtle">Subtle</option>
                   <option value="poster" data-text="poster">Poster</option>
+                  <option value="dynamic" data-text="dynamic">Dynamic</option>
                 </optgroup>
               </select>
               <select id="variant-select" tabindex="6">
