+++ src/styles/themes/dynamic.css
@@ -1,0 +1,40 @@
+[data-theme|="dynamic"] body {
+  position: relative;
+}
+
+[data-theme|="dynamic"] body:before {
+  content: "";
+  width: 10vw;
+  height: 10vw;
+  border-radius: 50%;
+  position: absolute;
+  left: calc(var(--actor-left) * 1%);
+  transition: left 60s;
+  transform: translateX(-50%);
+  top: 100px;
+}
+
+[data-theme|="dynamic"][data-actor="moon"] body:before {
+  background-color: rgb(189, 189, 189);
+}
+
+[data-theme|="dynamic"][data-scene="sunset"] {
+  --background-image: linear-gradient(
+    0deg,
+    rgba(173, 53, 25, 1) 0%,
+    rgba(250, 131, 73, 1) 9%,
+    rgba(170, 125, 131, 1) 21%,
+    rgba(50, 47, 76, 1) 65%,
+    rgba(28, 27, 43, 1) 86%
+  );
+  --accent-color: rgba(28, 27, 43, 1);
+}
+
+[data-theme|="dynamic"][data-scene="sunrise"] {
+  --background-image: linear-gradient(to right, #e0eafc, #cfdef3);
+}
+
+[data-theme|="dynamic"][data-scene="night"] {
+  --accent-color: rgb(189, 189, 189);
+  --background-image: linear-gradient(to bottom, #000120, #000000);
+}
